+++ eslint.config.js
@@ -1,0 +1,23 @@
+import js from "@eslint/js"
+import globals from "globals"
+
+export default [
+	js.configs.recommended,
+	{
+		languageOptions: {
+			ecmaVersion: 2022,
+			sourceType: "module",
+			globals: {
+				...globals.browser,
+				...globals.node,
+				Atomics: "readonly",
+				SharedArrayBuffer: "readonly",
+				DOMSize: "readonly"
+			},
+		},
+		rules: {
+			"no-prototype-builtins": "off",
+			"no-unused-vars": "off"
+		}
+	}
+];