+++ scripts/build-version-js.mjs
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import fs from "fs"
-import pkg from "../package.json" assert {type: "json"}
+import pkg from "../package.json" with {type: "json"}
 
 const src = `
 /**