+++ src/classes/extensions/StringExt.js
@@ -25,6 +25,74 @@ class StringExt extends Extension {
 	}
 */
 	/**
+	 * Converts string in ('PascalCase' | 'SNAKE_case' | 'KEBAB-case') notation to 'camelCase' string notation
+	 *
+	 * @param {string} sourceCase Source string notation, oneof(camel, pascal, snake, kebab)
+	 * @return {string} 'camelCase' notation value
+	 */
+	toCamelCase(sourceCase) {
+		switch (sourceCase) {
+			case "camel": return this;
+			case "pascal": return this.replace(/^./, c => c.toLowerCase());
+			case "snake": return this.toLowerCase().replace(/_([a-z])/g, (m, c) => c.toUpperCase()).replace(/^./, c => c.toLowerCase());
+			case "kebab": return this.toLowerCase().replace(/-([a-z])/g, (m, c) => c.toUpperCase()).replace(/^./, c => c.toLowerCase());
+
+			default: throw new Error(`Unsupported source case: ${sourceCase}, expected oneof(camel, pascal, snake, kebab)`);
+		}
+	}
+
+	/**
+	 * Converts string in ('camelCase' | 'SNAKE_case' | 'KEBAB-case') notation to 'PascalCase' string notation
+	 *
+	 * @param {string} sourceCase Source string notation, oneof(camel, pascal, snake, kebab)
+	 * @return {string} 'PascalCase' notation value
+	 */
+	toPascalCase(sourceCase) {
+		switch (sourceCase) {
+			case "pascal": return this;
+			case "camel": return this.replace(/^./, c => c.toUpperCase());
+			case "snake": return this.toLowerCase().replace(/_([a-z])/g, (m, c) => c.toUpperCase()).replace(/^./, c => c.toUpperCase());
+			case "kebab": return this.toLowerCase().replace(/-([a-z])/g, (m, c) => c.toUpperCase()).replace(/^./, c => c.toUpperCase());
+
+			default: throw new Error(`Unsupported source case: ${sourceCase}, expected oneof(camel, pascal, snake, kebab)`);
+		}
+	}
+
+	/**
+	 * Converts string in ('camelCase' | 'PascalCase' | 'KEBAB-case') notation to 'SNAKE_case' string notation
+	 *
+	 * @param {string} sourceCase Source string notation, oneof(camel, pascal, snake, kebab)
+	 * @return {string} 'SNAKE_case' notation value
+	 */
+	toSnakeCase(sourceCase) {
+		switch (sourceCase) {
+			case "snake": return this;
+			case "kebab": return this.replaceAll("-", "_");
+			case "camel": return this.replace(/[A-Z]/g, m => "_" + m);
+			case "pascal": return this.replace(/([a-z])([A-Z])/g, "$1_$2");
+
+			default: throw new Error(`Unsupported source case: ${sourceCase}, expected oneof(camel, pascal, snake, kebab)`);
+		}
+	}
+
+	/**
+	 * Converts string in ('camelCase' | 'PascalCase' | 'SNAKE_case') notation to 'KEBAB-case' string notation
+	 *
+	 * @param {string} sourceCase Source string notation, oneof(camel, pascal, snake, kebab)
+	 * @return {string} 'KEBAB-case' notation value
+	 */
+	toKebabCase(sourceCase) {
+		switch (sourceCase) {
+			case "kebab": return this;
+			case "snake": return this.replaceAll("_", "-");
+			case "camel": return this.replace(/([a-z])([A-Z])/g, "$1-$2");
+			case "pascal": return this.replace(/([a-z])([A-Z])/g, "$1-$2");
+
+			default: throw new Error(`Unsupported source case: ${sourceCase}, expected oneof(camel, pascal, snake, kebab)`);
+		}
+	}
+
+	/**
 	 * Converts string to char array
 	 *
 	 * @param {boolean} [bytes=false] Result type, When is true but string contains multi-bytes symbols is treated as false